feat(req 16): cria botão de carrinho com valor total do pedido e redireciona usuário para checkout
