+++ front-end/src/components/Chechout/TableItens.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function TableItens() {
+  return (
+    <h1>Checkout</h1>
+  );
+}
