+++ front-end/src/routes/routes.js
@@ -5,6 +5,7 @@ import { Route, useLocation } from 'react-router';
 import Login from '../pages/Login';
 import Register from '../pages/Register';
 import Products from '../pages/customer/Products';
+import TableItens from '../components/Chechout/TableItens';
 
 const Routes = () => {
   const location = useLocation();
@@ -14,10 +15,10 @@ const Routes = () => {
       {location.pathname === '/'
         ? <Redirect to="/login" />
         : null}
+      <Route exact path="/customer/checkout" component={ TableItens } />
       <Route exact path="/register" component={ Register } />
       <Route exact path="/customer/products" component={ Products } />
       <Route path="/" component={ Login } />
-
     </Switch>
   );
 };
